Raid fixes (#10509)
Rework raid messages for raid ending, they now got an individual finished message followed by a general victory message for the whole raid
Fix campfires getting spawned incorrectly, due to not loading a chunk when required to
Reworked North/South directions util, it is now split into 8 sections a 45° to give proper ranges for North/Northeast etc
Fix door break speed, research no longer strengthens normal doors and doors break slightly faster with more raiders hitting it.
Improved despawn of few remaining raiders to not stall raids for longer than needed
Full raid end is now handled by the raid manager
Added more message variants to raid end messages
Removed attackspeed scaling from archers, as those already scale in attack damage

+++ src/main/java/com/minecolonies/core/entity/mobs/aitasks/EntityAIBreakDoor.java
@@ -90,7 +90,7 @@ public class EntityAIBreakDoor extends BreakDoorGoal
         }
         else
         {
-            int fasterBreakPerXNearby = 5;
+            double fasterBreakPerXNearby = 5;
 
             if (mob instanceof AbstractEntityMinecoloniesRaider && !mob.level().isClientSide() && mob.level().getBlockState(doorPos).getBlock() instanceof AbstractBlockGate)
             {
